build scirpt
